refactor: split process unit tests
